+++ SpriteBatchDemo/SpriteBatchDemo/Classes/Item.cs
@@ -17,7 +17,7 @@ public class Item
         Position = SRandom.Vector2(320, 480);
         Angle = SRandom.Float( (float)Math.PI );
         Speed = SRandom.Float(200, 700);
-        Velocity = new Vector2((float)Math.Cos((double)Angle), (float)Math.Sin((double)Angle));
+        Velocity = new Vector2((float)Math.Cos(Angle), (float)Math.Sin(Angle));
         Color = Color.Multiply(Color.White, SRandom.Float(1));
         SpriteIndex = SRandom.Int();
         RectangleIndex = SRandom.Int();
